+++ src/lib.rs
@@ -149,22 +149,26 @@
 //! The resulting crate must provide an opt-in `rt` feature and depend on these crates:
 //!
 //! - [`critical-section`](https://crates.io/crates/critical-section) v1.x
-//! - [`riscv`](https://crates.io/crates/riscv) v0.9.x (if target is RISC-V)
-//! - [`riscv-rt`](https://crates.io/crates/riscv-rt) v0.9.x (if target is RISC-V)
+//! - [`riscv`](https://crates.io/crates/riscv) v0.9.x (if target is RISC-V) TODO update version
+//! - [`riscv-peripheral`](https://crates.io/crates/riscv-peripheral) v0.9.x (if target is RISC-V and has standard peripherals) TODO update version
+//! - [`riscv-rt`](https://crates.io/crates/riscv-rt) v0.9.x (if target is RISC-V) TODO update version
 //! - [`vcell`](https://crates.io/crates/vcell) v0.1.x
 //!
-//! The `*-rt` dependencies must be optional only enabled when the `rt` feature is enabled. The
-//! `Cargo.toml` of the device crate will look like this for a RISC-V target:
+//! The `*-rt` dependencies must be optional only enabled when the `rt` feature is enabled.
+//! If target is RISC-V and supports vectored mode, you must include a feature `v-trap` to activate `riscv-rt/v-trap`.
+//! The `Cargo.toml` of the device crate will look like this for a RISC-V target:
 //!
 //! ``` toml
 //! [dependencies]
 //! critical-section = { version = "1.0", optional = true }
-//! riscv = "0.9.0"
-//! riscv-rt = { version = "0.9.0", optional = true }
+//! riscv = "0.9.0" // TODO update version
+//! riscv-peripheral = "0.9.0" // TODO update version
+//! riscv-rt = { version = "0.9.0", optional = true } // TODO update version
 //! vcell = "0.1.0"
 //!
 //! [features]
 //! rt = ["riscv-rt"]
+//! v-trap = ["rt", "riscv-rt/v-trap"]
 //! ```
 //!
 //! # Peripheral API